Extract GitHubChangelogGenerator::FileParserChooser to its own file
