+++ lib/github_changelog_generator/parser_file.rb
@@ -1,31 +1,7 @@
 # frozen_string_literal: true
 
-require "pathname"
-
 module GitHubChangelogGenerator
   ParserError = Class.new(StandardError)
-
-  class FileParserChooser
-    def initialize(options)
-      @options     = options
-      @config_file = Pathname.new(options[:config_file])
-    end
-
-    def parse!(_argv)
-      return nil unless (path = resolve_path)
-
-      ParserFile.new(@options, File.open(path)).parse!
-    end
-
-    def resolve_path
-      return @config_file if @config_file.exist?
-
-      path = @config_file.expand_path
-      return path if File.exist?(path)
-
-      nil
-    end
-  end
 
   # ParserFile is a configuration file reader which sets options in the
   # given Hash.
